requirements.txt
django
django-restframework
cloudinary
python-dotenv
Pillow